+ DynamicGameObject has owner (game) & id is a short instead of int

+++ src/com/android/game/clash_of_the_balls/game/Game.java
@@ -74,7 +74,7 @@ public class Game extends GameBase implements UIBase {
 		m_view.applyView(renderer);
 		
 		m_game_field.draw(renderer);
-		for (Map.Entry<Integer, DynamicGameObject> entry : m_game_objects.entrySet()) {
+		for (Map.Entry<Short, DynamicGameObject> entry : m_game_objects.entrySet()) {
 			entry.getValue().draw(renderer);
 		}
